Before sending a SET_RES command over the bus, check if the sound card actually supports SET_RES for this value.
This prevents stalls on the USB bus. Normally this is not required
since stalls are just a "perfectly normal" thing, but the i.mx 5
EHCI driver does not seem to like those stalls and spews error
messages. Subsequently, the system is unstable and crashes.
Even though I cannot *prove* that this behaviour is related to
the error messages during the stalls, it is better to prevent the
stalls altogether.

Testing shows that the stability of the system has improved after
applying this patch.