Partially revert changes done in commit 43b12a00b80515d71fc1bab5fbb0d75360c40f1f
This seems to fix some kernel crashes.

No idea why we need to keep the last free td instead of freeing it directly.

Fixes
 > ffmpeg -i [some mp3 stream] -t 60:00 -acodec copy /some/mp3/file

This command was crshing always when ffmpeg was connecting to the
internet via the "g_ether" network gagdet driver.

The crash is now fixed by this commit.