USB OTG power management fixes.
Seperate the OTG suspend/resume routines from the PM suspend routines
to avoid problems with concurrent function calls to these methods.

This means that it is now possible to plug/unplug the USB cable
while the Kindle is suspended.

Before that, we sometimes misdetected PM resume requests for OTG
resume requests. Thus, the host and gadget drivers may resume
the USB bus even though this was not requested by the OTG state
machine.