[USB] Undo commit 81a4b7af4abcf4b28c7dc549fcf3a4ae3f4dd65b
It appears that the file storage gadget hangs when ->disconnect()
is called immediatly before going into LPM.

The fix is not to call ->disconnect() at all. Well, technically we
*should* call the method, but in Documentation/DocBook/gadget.tmpl
is says:
"Some devices can't detect disconnect, so this might not be called
except as part of controller shutdown."

So it seems like USB drivers don't *have* to call ->disconnect()
on every USB disconnect.